+++ src/playlist-loader.js
@@ -21,6 +21,7 @@ import {
 import {getKnownPartCount} from './playlist.js';
 import {merge} from './util/vjs-compat';
 import DateRangesStorage from './util/date-ranges';
+import { getStreamingNetworkErrorMetadata } from './error-codes.js';
 
 const { EventTarget } = videojs;
 
@@ -371,6 +372,34 @@ export const refreshDelay = (media, update) => {
   return (media.partTargetDuration || media.targetDuration || 10) * 500;
 };
 
+const playlistMetadataPayload = (playlists, type, isLive) => {
+  if (!playlists) {
+    return;
+  }
+  const renditions = [];
+
+  playlists.forEach((playlist) => {
+    // we need attributes to populate rendition data.
+    if (!playlist.attributes) {
+      return;
+    }
+    const { BANDWIDTH, RESOLUTION, CODECS } = playlist.attributes;
+
+    renditions.push({
+      id: playlist.id,
+      bandwidth: BANDWIDTH,
+      resolution: RESOLUTION,
+      codecs: CODECS
+    });
+  });
+
+  return {
+    type,
+    isLive,
+    renditions
+  };
+};
+
 /**
  * Load a playlist from a remote location
  *
@@ -486,23 +515,29 @@ export default class PlaylistLoader extends EventTarget {
       message: `HLS playlist request error at URL: ${uri}.`,
       responseText: xhr.responseText,
       code: (xhr.status >= 500) ? 4 : 2,
-      metadata: {
-        errorType: videojs.Error.HlsPlaylistRequestError
-      }
+      metadata: getStreamingNetworkErrorMetadata({ requestType: xhr.requestType, request: xhr, error: xhr.error })
     };
 
     this.trigger('error');
   }
 
   parseManifest_({url, manifestString}) {
-    return parseManifest({
-      onwarn: ({message}) => this.logger_(`m3u8-parser warn for ${url}: ${message}`),
-      oninfo: ({message}) => this.logger_(`m3u8-parser info for ${url}: ${message}`),
-      manifestString,
-      customTagParsers: this.customTagParsers,
-      customTagMappers: this.customTagMappers,
-      llhls: this.llhls
-    });
+    try {
+      return parseManifest({
+        onwarn: ({message}) => this.logger_(`m3u8-parser warn for ${url}: ${message}`),
+        oninfo: ({message}) => this.logger_(`m3u8-parser info for ${url}: ${message}`),
+        manifestString,
+        customTagParsers: this.customTagParsers,
+        customTagMappers: this.customTagMappers,
+        llhls: this.llhls
+      });
+    } catch (error) {
+      this.error = error;
+      this.error.metadata = {
+        errorType: videojs.Error.StreamingHlsPlaylistParserError,
+        error
+      };
+    }
   }
 
   /**
@@ -522,6 +557,14 @@ export default class PlaylistLoader extends EventTarget {
     this.request = null;
     this.state = 'HAVE_METADATA';
 
+    const metadata = {
+      playlistInfo: {
+        type: 'media',
+        uri: url
+      }
+    };
+
+    this.trigger({type: 'playlistparsestart', metadata });
     const playlist = playlistObject || this.parseManifest_({
       url,
       manifestString: playlistString
@@ -550,7 +593,8 @@ export default class PlaylistLoader extends EventTarget {
     }
 
     this.updateMediaUpdateTimeout_(refreshDelay(this.media(), !!update));
-
+    metadata.parsedPlaylist = playlistMetadataPayload(this.main.playlists, metadata.playlistInfo.type, !this.media_.endList);
+    this.trigger({ type: 'playlistparsecomplete', metadata });
     this.trigger('loadedplaylist');
   }
 
@@ -690,6 +734,14 @@ export default class PlaylistLoader extends EventTarget {
     }
 
     this.pendingMedia_ = playlist;
+    const metadata = {
+      playlistInfo: {
+        type: 'media',
+        uri: playlist.uri
+      }
+    };
+
+    this.trigger({ type: 'playlistrequeststart', metadata });
 
     this.request = this.vhs_.xhr({
       uri: playlist.resolvedUri,
@@ -708,6 +760,8 @@ export default class PlaylistLoader extends EventTarget {
       if (error) {
         return this.playlistRequestError(this.request, playlist, startingState);
       }
+
+      this.trigger({ type: 'playlistrequestcomplete', metadata });
 
       this.haveMetadata({
         playlistString: req.responseText,
@@ -836,7 +890,14 @@ export default class PlaylistLoader extends EventTarget {
       }, 0);
       return;
     }
+    const metadata = {
+      playlistInfo: {
+        type: 'multivariant',
+        uri: this.src
+      }
+    };
 
+    this.trigger({ type: 'playlistrequeststart', metadata });
     // request the specified URL
     this.request = this.vhs_.xhr({
       uri: this.src,
@@ -858,22 +919,26 @@ export default class PlaylistLoader extends EventTarget {
           responseText: req.responseText,
           // MEDIA_ERR_NETWORK
           code: 2,
-          metadata: {
-            errorType: videojs.Error.HlsPlaylistRequestError
-          }
+          metadata: getStreamingNetworkErrorMetadata({ requestType: req.requestType, request: req, error })
         };
         if (this.state === 'HAVE_NOTHING') {
           this.started = false;
         }
         return this.trigger('error');
       }
+      this.trigger({ type: 'playlistrequestcomplete', metadata });
 
       this.src = resolveManifestRedirect(this.src, req);
 
+      this.trigger({ type: 'playlistparsestart', metadata });
       const manifest = this.parseManifest_({
         manifestString: req.responseText,
         url: this.src
       });
+
+      // we haven't loaded any variant playlists here so we default to false for isLive.
+      metadata.parsedPlaylist = playlistMetadataPayload(manifest.playlists, metadata.playlistInfo.type, false);
+      this.trigger({ type: 'playlistparsecomplete', metadata });
 
       this.setupInitialPlaylist(manifest);
     });