+++ test/dash-playlist-loader.test.js
@@ -723,10 +723,6 @@ QUnit.test('addSidxSegments_: adds/triggers error on invalid container', functio
     code: 2,
     internal: true,
     message: 'Unsupported unknown container type for sidx segment at URL: sidx.mp4',
-    metadata: {
-      errorType: 'unsupported-sidx-container-error',
-      sidxContainer: 'unknown'
-    },
     playlist,
     response: '',
     status: 200
@@ -2026,7 +2022,13 @@ QUnit.test('addSidxSegments_: errors if request for sidx fails', function(assert
     {
       status: 500,
       message: 'DASH request error at URL: sidx.mp4',
-      metadata: undefined,
+      metadata: {
+        errorType: 'networkbadstatus',
+        headers: {},
+        requestType: 'dash-sidx',
+        status: 500,
+        uri: 'sidx.mp4'
+      },
       response: '',
       code: 2
     },